requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.0.7
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-api-key==2.3.0
djangorestframework-xml==2.0.0
Markdown==3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
